@@ -8,35 +8,33 @@
     "generate": "nuxt generate",
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare",
-    "lint:js": "eslint --ext \".ts,.vue\" --ignore-path .gitignore .",
-    "lint:prettier": "prettier --check .",
-    "lint": "pnpm lint:js && pnpm lint:prettier",
-    "lint:fix": "prettier --write --list-different . && pnpm lint:js --fix"
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
   "devDependencies": {
-    "@nuxt/devtools": "^0.6.4",
-    "@nuxtjs/eslint-config-typescript": "^12.0.0",
+    "@nuxt/eslint": "^0.5.7",
+    "@nuxt/image": "^1.8.1",
+    "@nuxt/ui": "^2.18.6",
     "@nuxtjs/tailwindcss": "^6.6.7",
     "@types/node": "^18",
-    "@typescript-eslint/parser": "^5.59.2",
     "@vueuse/nuxt": "^10.1.2",
-    "eslint": "^8.39.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "nuxt": "^3.4.3",
+    "nuxt": "^3.13.2",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.8",
-    "typescript": "^5.0.4",
+    "typescript": "^5.6.2",
     "vite": "^4.3.5",
     "vue-inline-svg": "^3.1.2"
   },
   "dependencies": {
-    "@headlessui/tailwindcss": "^0.1.3",
+    "@iconify-json/mdi": "^1.2.0",
     "@tanstack/vue-query": "^4.29.15",
     "@vite-pwa/nuxt": "^0.0.7",
     "@vuelidate/core": "^2.0.2",
     "@vuelidate/validators": "^2.0.2",
-    "nuxt-headlessui": "^1.1.4",
-    "qs": "^6.11.2"
+    "chart.js": "^4.4.4",
+    "date-fns": "^4.1.0",
+    "qs": "^6.11.2",
+    "v-calendar": "^3.1.2",
+    "vue-chartjs": "^5.3.1"
   }
 }
