@@ -1,7 +1,11 @@
-module.exports = {
-  content: [],
+import type { Config } from 'tailwindcss'
+
+export default <Partial<Config>>{
   theme: {
-    extend: {},
-  },
-  plugins: [require("@headlessui/tailwindcss")],
-};
+    extend: {
+      colors: {
+        
+      }
+    }
+}
+}
