@@ -1,9 +1,9 @@
 <template>
-  <form class="mt-5 flex flex-col gap-3" @submit.prevent="submitCharge">
+  <UForm class="mt-5 flex flex-col gap-3" :state="chargeState" @submit="submitCharge">
     <div>
-      <input
+      <UInput
         v-model="v$.name.$model"
-        class="text-md mb-2 block w-full rounded-md bg-gray-100 px-5 py-2 invalid:border-red-300 focus:outline-orange-300"
+        size="lg"
         placeholder="Nom"
         type="text"
       />
@@ -12,25 +12,27 @@
       </span>
     </div>
     <div>
-      <Multiselect v-model="v$.bank.$model" :items="banks"></Multiselect>
+      <UInputMenu v-model="v$.bank.$model" size="lg" :options="banks" option-attribute="name"/>
       <span v-if="v$.bank.$error" class="text-red-600">
         {{ v$.bank.$errors[0]?.$message }}
       </span>
     </div>
     <div>
-      <Multiselect
+      <UInputMenu
         v-model="v$.chargeType.$model"
-        placeholder="Banque"
-        :items="chargeTypes"
-      ></Multiselect>
+        size="lg"
+        placeholder="Type"
+        :options="chargeTypes"
+        option-attribute="name"
+      />
       <span v-if="v$.chargeType.$error" class="text-red-600">
         {{ v$.chargeType.$errors[0]?.$message }}
       </span>
     </div>
     <div>
-      <input
+      <UInput
         v-model="v$.date.$model"
-        class="text-md mb-2 block w-full rounded-md bg-gray-100 px-5 py-2 invalid:border-red-300 focus:outline-orange-300"
+        size="lg"
         placeholder="Date"
         type="date"
       />
@@ -39,10 +41,11 @@
       </span>
     </div>
     <div>
-      <input
-        v-model="v$.amount.$model"
+      <UInput
+      v-model="v$.amount.$model"
+        trailing-icon="i-mdi-euro"
+        size="lg"
         step="0.01"
-        class="text-md mb-2 block w-full rounded-md bg-gray-100 px-5 py-2 invalid:border-red-300 focus:outline-orange-300"
         placeholder="Montant (100€)"
         type="number"
       />
@@ -51,33 +54,31 @@
       </span>
     </div>
     <div class="flex w-full items-center justify-center">
-      <button
+      <UButton
         type="submit"
         class="rounded-md bg-green-400 px-5 py-2 text-white hover:bg-green-500 focus:outline-green-600"
-      >
-        Enregistrer
-      </button>
+        label="Enregistrer"
+      />
     </div>
-  </form>
+  </UForm>
 </template>
 
 <script setup lang="ts">
-import { minValue, required } from "@vuelidate/validators";
+import type { Charge } from "@type/Charge";
 import useVuelidate from "@vuelidate/core";
+import { minValue, required } from "@vuelidate/validators";
 import { computed, reactive } from "vue";
-import { Bank } from "../@type/Bank";
-import { useApiRoutes, getMembers } from "../composables/api";
-import { Charge } from "../@type/Charge";
-import Multiselect from "./Multiselect.vue";
+import { getMembers, useApiRoutes } from "../composables/api";
+import { format } from "date-fns";
 
 const api = useApiRoutes();
 
-const props = defineProps<{
+const props = defineProps<{ 
   initialValue?: Charge;
 }>();
 
 const emit = defineEmits<{
-  onSubmit: [value: any];
+  onSubmit: [value];
 }>();
 
 const currentUser = ref();
@@ -104,11 +105,11 @@ const { data: chargeTypes } = useAsyncData("chargeTypes", () => {
 
 const chargeState = reactive({
   name: props.initialValue?.name || "",
-  amount: props.initialValue?.amount || "",
+  amount: props.initialValue?.amount || 0,
   state: props.initialValue?.state || false,
   bank: props.initialValue?.bank || banks.value?.[0],
   chargeType: props.initialValue?.chargeType || chargeTypes.value?.[0],
-  date: props.initialValue?.date || "",
+  date: props.initialValue?.date ? format(new Date(props.initialValue?.date), 'yyyy-MM-dd') : '',
 });
 const v$ = useVuelidate(chargeRules, chargeState);
 
@@ -118,9 +119,9 @@ const submitCharge = async () => {
     return;
   }
 
-  const charge = {
+  const charge: Charge = {
     name: v$.value.name.$model,
-    amount: v$.value.amount.$model.toString(),
+    amount: v$.value.amount.$model,
     state: v$.value.state.$model,
     date: v$.value.date.$model,
     bank: v$.value.bank.$model?.["@id"],
