@@ -1,16 +1,23 @@
 <template>
   <section class="container mx-auto grid grid-cols-1 gap-5 p-6 lg:grid-cols-2">
-    <h1 class="col-span-full text-xl font-bold text-gray-800">
-      Bonjour, <br />
-      <span class="capitalize text-orange-300">{{
-        currentUser?.userIdentifier || "Utilisateur"
-      }}</span>
-    </h1>
+    <div class="col-span-full flex justify-between text-xl font-bold text-gray-800">
+      <h1>Bonjour, <br >
+        <span class="capitalize text-orange-300">{{
+          currentUser?.userIdentifier || "Utilisateur"
+        }}</span>
+      </h1>
+      <div>
+        <UButton icon="i-mdi-logout" variant="ghost" label="Deconnexion" @click="signOut"/>
+      </div>
+    </div>
     <div class="col-span-1 flex flex-col rounded-lg bg-blue-500 p-5 text-white">
       <span>Restant ce mois-ci</span>
       <span class="text-2xl font-extrabold">{{ sumCurrent || 0 }}€</span>
       <span class="self-end">Total</span>
       <span class="self-end text-xl font-extrabold">{{ sum || 0 }}€</span>
+    </div>
+    <div v-if="charges?.length > 0" class="col-span-1">
+        <ChargeCategoryChart :charges="charges" />
     </div>
     <div class="col-span-1">
       <div class="flex justify-between">
@@ -18,33 +25,14 @@
       </div>
       <div class="w-full overflow-auto py-5">
         <div class="flex w-fit gap-5">
-          <button
-            v-for="bank in banks"
-            :key="bank.id"
-            :title="bank.name"
-            class="aspect-square w-14 rounded-lg bg-gray-200 p-3 hover:bg-gray-300 focus:outline-orange-300"
-            :class="[
-              selectedBank === bank.id ? 'border-2 border-orange-300' : '',
-            ]"
-            @click.prevent="updateQuery(bank.id)"
-          >
+          <UButton v-for="bank in banks" :key="bank.id" size="xl" color="gray" :class="[selectedBank === bank.id ? 'border-2 border-offset-2 border-orange-500 dark:border-orange-400': '']" :title="bank.name" @click.prevent="updateQuery(bank.id)">
             <img
-              class="h-full w-full rounded-lg object-cover"
-              :src="bank.image"
-              :alt="bank.abbreviation"
-            />
-          </button>
-          <button
-            title="Tous"
-            class="aspect-square w-14 rounded-lg bg-gray-200 p-3 hover:bg-gray-300 focus:outline-orange-300"
-            :class="[!selectedBank ? 'border-2 border-orange-300' : '']"
-            @click.prevent="updateQuery()"
-          >
-            <InlineSvg
-              class="h-full w-full text-gray-800"
-              src="/svg/all.svg"
-            ></InlineSvg>
-          </button>
+            class="h-full w-full rounded-lg object-cover"
+            :src="bank.image"
+            :alt="bank.abbreviation"
+            >
+          </UButton>
+          <UButton icon="i-mdi-view-grid" size="xl" color="gray" :class="[!selectedBank ? 'border-2 border-offset-2 border-orange-500 dark:border-orange-400': '']" title="Tous" @click.prevent="updateQuery()"/>
         </div>
       </div>
     </div>
@@ -52,51 +40,45 @@
       <div class="flex justify-between">
         <h1 class="text-xl font-bold text-gray-300">Prélevements</h1>
         <div class="flex gap-2">
-          <button
-            title="Réinitiliser"
-            class="flex h-7 w-7 items-center justify-center rounded bg-orange-300 p-0 text-3xl font-extrabold text-white hover:bg-orange-400 focus:outline-orange-400"
-            @click.prevent="resetCharges"
-          >
-            <InlineSvg src="/svg/arrow.svg" class="w-ful h-full"></InlineSvg>
-          </button>
-          <button
-            title="Ajouter"
-            class="flex h-7 w-7 items-center justify-center rounded bg-green-300 p-0 text-3xl font-extrabold text-white hover:bg-green-400 focus:outline-green-400"
-            @click.prevent="openModal"
-          >
-            <InlineSvg
-              src="/svg/plus.svg"
-              class="h-full w-full text-white"
-            ></InlineSvg>
-          </button>
+          <UButton
+          title="Réinitiliser"
+          icon="i-mdi-reload"
+          color="orange"
+          @click="resetCharges"
+          />
+          <UButton title="Ajouter" icon="i-mdi-plus" @click="openModal"/>
         </div>
       </div>
       <ChargeCard
-        v-for="charge in charges"
-        :key="charge['@id']"
-        :charge="charge"
-        @updateState="updateChargeState"
-        @deleteCharge="deleteCharge"
-        @updateCharge="updateCharge"
+      v-for="charge in charges"
+      :key="charge['@id']"
+      :charge="charge"
+      @update-state="updateChargeState"
+      @delete-charge="deleteCharge"
+      @update-charge="updateCharge"
       />
     </div>
-    <ModalBottomSheet :is-open="isOpen" @close="closeModal">
-      <template #title>Ajouter un prélèvement</template>
-      <template #body>
-        <ChargeForm @onSubmit="submitCharge"></ChargeForm>
-      </template>
-    </ModalBottomSheet>
+    <teleport to="body">
+      <UModal v-model="isOpen">
+        <UCard>
+          <template #header>
+            <h3 class="text-center text-lg font-medium leading-6 text-orange-300">
+              Ajouter un prélèvement
+            </h3>
+          </template>
+          <ChargeForm @on-submit="submitCharge"/>
+        </UCard>
+      </UModal>
+    </teleport>
   </section>
 </template>
 <script lang="ts" setup>
-import { ref } from "vue";
-import InlineSvg from "vue-inline-svg";
-import { useApiRoutes, getMembers } from "../composables/api";
-import { ChargeType } from "../@type/ChargeType";
-import { Bank } from "../@type/Bank";
-import { Charge } from "../@type/Charge";
+import type { Bank } from "@type/Bank";
+import type { Charge } from "@type/Charge";
+import { computed, ref } from "vue";
+import ChargeCategoryChart from "../components/ChargeCategoryChart.vue";
 import ChargeForm from "../components/ChargeForm.vue";
-import ModalBottomSheet from "../components/ModalBottomSheet.vue";
+import { getMembers, useApiRoutes } from "../composables/api";
 
 definePageMeta({
   middleware: ["auth"],
@@ -105,43 +87,36 @@ definePageMeta({
 const currentUser = ref();
 
 const { authUser } = useAuthState();
+const { signOut } = useAuth()
 const api = useApiRoutes();
 
 currentUser.value = authUser.value;
 
 const sum = computed(() => {
   return charges.value
-    ?.reduce((partialSum: number, item: Charge) => partialSum + item.amount, 0)
-    .toFixed(2);
+  ?.reduce((partialSum: number, item: Charge) => partialSum + item.amount, 0)
+  .toFixed(2);
 });
 
 const sumCurrent = computed(() => {
   return charges.value
-    ?.reduce(
-      (partialSum: number, item: Charge) =>
-        partialSum + (!item.state ? item.amount : 0),
-      0
-    )
-    .toFixed(2);
+  ?.reduce(
+  (partialSum: number, item: Charge) =>
+  partialSum + (!item.state ? item.amount : 0),
+  0
+  )
+  .toFixed(2);
 });
 
-const { data: banks } = useAsyncData<never, never, Bank[]>("userBanks", () => {
+const { data: banks } = useAsyncData<Bank[]>("userBanks", () => {
   return getMembers(api.banks.getUserBanks());
 });
-
-const { data: chargeTypes } = useAsyncData<never, never, ChargeType[]>(
-  "chargeTypes",
-  () => {
-    return getMembers(api.chargeTypes.getCollection());
-  }
-);
 
 const query = ref({ "bank.id": banks.value?.[0]?.id });
 const {
   data: charges,
-  error,
-  refresh: refreshCharges,
-} = useAsyncData<never, never, Charge[]>("charges", () => {
+  refresh: refreshCharges
+} = useAsyncData<Charge[]>("charges", () => {
   return getMembers(api.charges.getCollection(query.value));
 });
 
@@ -157,6 +132,7 @@ const closeModal = () => {
 };
 
 const updateQuery = async (id?: number) => {
+  console.log('dedans');
   if (!id) {
     selectedBank.value = null;
     query.value = {};
@@ -167,7 +143,7 @@ const updateQuery = async (id?: number) => {
   await refreshCharges();
 };
 
-const submitCharge = async (charge: any) => {
+const submitCharge = async (charge: Ref<Charge>) => {
   try {
     await api.charges.create(charge.value);
     closeModal();
