@@ -6,9 +6,10 @@
       <div
         class="h-[60vh] w-full rounded-t-xl bg-white shadow dark:border dark:border-gray-700 dark:bg-gray-800 sm:max-w-md md:mt-0 md:h-fit md:rounded-md xl:p-0"
       >
-        <form
+        <UForm
+          :state="v$"
           class="space-y-4 p-6 sm:p-8 md:space-y-6"
-          @submit.prevent="submitForm"
+          @submit="submitForm"
         >
           <h1
             class="text-xl font-bold leading-tight tracking-tight text-orange-300 dark:text-white"
@@ -16,10 +17,9 @@
             Identifiez vous
           </h1>
           <div>
-            <input
+            <UInput
               v-model="v$.username.$model"
               :required="true"
-              class="text-md mb-2 block w-full rounded-md bg-gray-100 px-5 py-2 invalid:border-red-300 invalid:outline-red-500 focus:outline-orange-300"
               placeholder="Nom"
             />
             <span v-if="v$.username.$error" class="text-red-600">
@@ -27,26 +27,19 @@
             </span>
           </div>
           <div>
-            <input
-              v-model="v$.password.$model"
-              class="text-md mb-2 block w-full rounded-md bg-gray-100 px-5 py-2 invalid:border-red-300 focus:outline-orange-300"
-              placeholder="Mot de passe"
-              type="password"
-            />
+            <UInput v-model="v$.password.$model" placeholder="Mot de passe" type="password"/>
             <span v-if="v$.password.$error" class="text-red-600">
               {{ v$.password.$errors[0]?.$message }}
             </span>
           </div>
           <span v-if="errorLogin" class="text-red-600">{{ errorLogin }}</span>
           <div class="flex w-full items-center justify-center">
-            <button
+            <UButton
               type="submit"
-              class="rounded-md bg-green-400 px-5 py-2 text-white hover:bg-green-500 focus:outline-green-600"
-            >
-              Connexion
-            </button>
+              label="Connexion"
+            />
           </div>
-        </form>
+        </UForm>
       </div>
     </div>
   </section>
